# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Student Course API",
    "version": "1.0.0",
    "description": "API REST pour gérer les étudiants, les cours et les inscriptions."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/courses": {
      "get": {
        "summary": "Liste les cours",
        "responses": {
          "200": {
            "description": "Liste des cours",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Course" }
                }
              }
            }
          }
        }
      },

      "post": {
        "summary": "Crée un cours",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CourseInput" } }
          }
        },
        "responses": {
          "201": { "description": "Cours créé" },
          "400": { "description": "Titre déjà utilisé" }
        }
      }
    },
    "/courses/{id}": {
      "get": {
        "summary": "Récupère un cours et ses étudiants inscrits",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Cours trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "course": { "$ref": "#/components/schemas/Course" },
                    "students": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Student" }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Cours introuvable" }
        }
      },

      "put": {
        "summary": "Met à jour un cours",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CourseInput" } }
          }
        },
        "responses": {
          "200": { "description": "Cours mis à jour" },
          "400": { "description": "Titre déjà utilisé" },
          "404": { "description": "Cours introuvable" }
        }
      },

      "delete": {
        "summary": "Supprime un cours",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "Cours supprimé" },
          "400": { "description": "Impossible (étudiants inscrits)" },
          "404": { "description": "Cours introuvable" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Student": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string" }
        }
      },
...
      "Course": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "teacher": { "type": "string" }
        }
      },
      "CourseInput": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "teacher": { "type": "string" }
        },
        "required": ["title", "teacher"]
      }
    }
  }
}